blogComp added
