+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/peterwaller/Desktop/FinalProject/treasures-web-frontend/src/components/Footer.js":"1","/Users/peterwaller/Desktop/FinalProject/treasures-web-frontend/src/pages/blogs/blogs.js":"2"},{"size":985,"mtime":1606819490487,"results":"3","hashOfConfig":"4"},{"size":1974,"mtime":1606820392649,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1u1osyw",{"filePath":"8","messages":"9","errorCount":0,"warningCount":8,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/peterwaller/Desktop/FinalProject/treasures-web-frontend/src/components/Footer.js",[],"/Users/peterwaller/Desktop/FinalProject/treasures-web-frontend/src/pages/blogs/blogs.js",["10","11","12","13","14","15","16","17"],{"ruleId":"18","severity":1,"message":"19","line":2,"column":27,"nodeType":"20","messageId":"21","endLine":2,"endColumn":33},{"ruleId":"18","severity":1,"message":"22","line":2,"column":35,"nodeType":"20","messageId":"21","endLine":2,"endColumn":40},{"ruleId":"18","severity":1,"message":"23","line":2,"column":42,"nodeType":"20","messageId":"21","endLine":2,"endColumn":46},{"ruleId":"18","severity":1,"message":"24","line":7,"column":8,"nodeType":"20","messageId":"21","endLine":7,"endColumn":14},{"ruleId":"18","severity":1,"message":"25","line":7,"column":16,"nodeType":"20","messageId":"21","endLine":7,"endColumn":25},{"ruleId":"18","severity":1,"message":"26","line":8,"column":8,"nodeType":"20","messageId":"21","endLine":8,"endColumn":13},{"ruleId":"18","severity":1,"message":"27","line":9,"column":8,"nodeType":"20","messageId":"21","endLine":9,"endColumn":12},{"ruleId":"18","severity":1,"message":"28","line":9,"column":14,"nodeType":"20","messageId":"21","endLine":9,"endColumn":21},"no-unused-vars","'Router' is defined but never used.","Identifier","unusedVar","'Route' is defined but never used.","'Link' is defined but never used.","'isEdit' is assigned a value but never used.","'setIsEdit' is assigned a value but never used.","'blogs' is assigned a value but never used.","'blog' is assigned a value but never used.","'setBlog' is assigned a value but never used."]
+[{"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/index.js":"1","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/contact/contact.js":"2","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/blogs.js":"3","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/declaration/declaration.js":"4","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/hearts/hearts.js":"5","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/resources/resources.js":"6","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/BlogList.js":"7","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/blogComp.js":"8","/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/components/Footer.js":"9"},{"size":1609,"mtime":1606924794432,"results":"10","hashOfConfig":"11"},{"size":1535,"mtime":1606924794433,"results":"12","hashOfConfig":"11"},{"size":316,"mtime":1606928921958,"results":"13","hashOfConfig":"11"},{"size":288,"mtime":1606924794433,"results":"14","hashOfConfig":"11"},{"size":1196,"mtime":1606924794434,"results":"15","hashOfConfig":"11"},{"size":273,"mtime":1606924794435,"results":"16","hashOfConfig":"11"},{"size":935,"mtime":1606926628546,"results":"17","hashOfConfig":"11"},{"size":1940,"mtime":1606929851823,"results":"18","hashOfConfig":"11"},{"size":1198,"mtime":1606924794432,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7j2d2z",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/index.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/contact/contact.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/blogs.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/declaration/declaration.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/hearts/hearts.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/resources/resources.js",[],["39","40"],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/BlogList.js",[],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/pages/blogs/blogComp.js",["41","42","43"],"/Users/kristinterrazas/Desktop/treasures/treasures-web-frontend/src/components/Footer.js",[],{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","replacedBy":"47"},{"ruleId":"48","severity":1,"message":"49","line":6,"column":10,"nodeType":"50","messageId":"51","endLine":6,"endColumn":15},{"ruleId":"48","severity":1,"message":"52","line":7,"column":10,"nodeType":"50","messageId":"51","endLine":7,"endColumn":14},{"ruleId":"48","severity":1,"message":"53","line":7,"column":16,"nodeType":"50","messageId":"51","endLine":7,"endColumn":23},"no-native-reassign",["54"],"no-negated-in-lhs",["55"],"no-unused-vars","'blogs' is assigned a value but never used.","Identifier","unusedVar","'blog' is assigned a value but never used.","'setBlog' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
